requirements/local.txt
# Local development dependencies go here
-r base.txt
-e git+git://github.com/dcramer/django-devserver#egg=django-devserver
behave==1.2.5
coverage==3.7.1
django-behave==0.1.4
django-debug-toolbar-line-profiler==0.4.0
django-debug-toolbar-template-profiler==1.0.1
django-debug-toolbar-template-timings==0.6.4
django-debug-toolbar==1.3.0
django-discover-runner==1.0
django-nose==1.3
enum34==1.0.4
line-profiler==1.0
nose==1.3.6
parse-type==0.3.4
parse==1.6.6
selenium==2.45.0
Sphinx==1.3.1
splinter==0.7.2
yanc==0.3.3
